+++ packages/core/src/types/on-queue-success.type.ts
@@ -1,7 +1,5 @@
-import { TestFile } from '../models/index.js';
-
 export type onQueueSuccessType = (paths: string[]) => Promise<{
-  recordedPaths: TestFile[] | Record<string, number>;
-  failedPaths: string[];
+  recordedPaths: Record<string, number>;
+  failedPaths: Set<string>;
   isTestSuiteGreen: boolean;
 }>;