+++ packages/cypress/__tests__/knapsack-pro-cypress.spec.ts
@@ -27,21 +27,4 @@ describe('#normalizePaths', () => {
 
     expect(actual).toEqual(expected);
   });
-
-  it('transforms linePaths to filePaths', () => {
-    const scheduledPaths = ['x.spec.js:1', 'y.spec.js'];
-
-    const recordedPaths = {
-      'x.spec.js:1': 4,
-      'y.spec.js:1': 3,
-    };
-
-    const actual = normalizePaths(scheduledPaths, recordedPaths);
-    const expected = {
-      'x.spec.js:1': 4,
-      'y.spec.js': 3,
-    };
-
-    expect(actual).toEqual(expected);
-  });
 });