+++ packages/jest/src/paths-finder.ts
@@ -1,31 +1,30 @@
 import { glob } from 'glob';
 import { minimatch } from 'minimatch';
 
-import { KnapsackProLogger, TestFile } from '@knapsack-pro/core';
+import { KnapsackProLogger } from '@knapsack-pro/core';
 import { EnvConfig } from './env-config.js';
 import * as Urls from './urls.js';
 
-export class TestFilesFinder {
-  public static allTestFiles(): TestFile[] {
-    const testFiles = glob
+export class PathsFinder {
+  public static allPaths(): string[] {
+    const paths = glob
       .sync(EnvConfig.testFilePattern)
-      .filter((testFilePath: string) => {
+      .filter((path: string) => {
         if (EnvConfig.testFileExcludePattern) {
-          return !minimatch(testFilePath, EnvConfig.testFileExcludePattern, {
+          return !minimatch(path, EnvConfig.testFileExcludePattern, {
             matchBase: true,
           });
         }
         return true;
       })
       .filter(
-        (testFilePath: string) =>
+        (path: string) =>
           // ignore test file paths inside node_modules because it's default Jest behavior
           // https://jestjs.io/docs/en/22.2/configuration#testpathignorepatterns-array-string
-          !testFilePath.match(/node_modules/),
-      )
-      .map((testFilePath: string) => ({ path: testFilePath }));
+          !path.match(/node_modules/),
+      );
 
-    if (testFiles.length === 0) {
+    if (paths.length === 0) {
       const knapsackProLogger = new KnapsackProLogger();
 
       const errorMessage = `Test files cannot be found.\nPlease set KNAPSACK_PRO_TEST_FILE_PATTERN matching your test directory structure.\nLearn more: ${Urls.NO_TESTS_FOUND}`;
@@ -34,6 +33,6 @@ export class TestFilesFinder {
       throw errorMessage;
     }
 
-    return testFiles;
+    return paths;
   }
 }