+++ packages/playwright-example-test-suite/tests/serial.spec.ts
@@ -1,0 +1,22 @@
+import { test, type Page } from '@playwright/test';
+
+// Annotate entire file as serial.
+test.describe.configure({ mode: 'serial' });
+
+let page: Page;
+
+test.beforeAll(async ({ browser }) => {
+  page = await browser.newPage();
+});
+
+test.afterAll(async () => {
+  await page.close();
+});
+
+test('runs first', async () => {
+  await page.goto('https://playwright.dev/');
+});
+
+test('runs second', async () => {
+  await page.getByText('Get Started').click();
+});